dubbo-samples-notify:change groupId and package to org.apache.dubbo
