+++ dubbo-samples-notify/src/main/java/org/apache/dubbo/samples/notify/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.notify;
+package org.apache.dubbo.samples.notify;
 
 import java.io.File;
 import java.lang.reflect.Method;
