+++ dubbo-samples-notify/src/main/java/org/apache/dubbo/samples/notify/impl/NotifyImpl.java
@@ -17,14 +17,24 @@
  *
  */
 
-package com.alibaba.dubbo.samples.notify.impl;
+package org.apache.dubbo.samples.notify.impl;
 
-import com.alibaba.dubbo.samples.notify.api.DemoService;
+import java.util.HashMap;
+import java.util.Map;
 
+import org.apache.dubbo.samples.notify.api.Notify;
 
-public class DemoServiceImpl implements DemoService{
+public class NotifyImpl implements Notify {
 
-    public String sayHello(int id) {
-        return "aaa";
+    public Map<Integer, String> ret = new HashMap<Integer, String>();
+    @Override
+    public void onreturn(String name, int id) {
+        ret.put(id, name);
+        System.out.println("onreturn: " + name);
+    }
+
+    @Override
+    public void onthrow(Throwable ex, String name, int id) {
+        System.out.println("onthrow: " + name);
     }
 }
