+++ dubbo-samples-notify/src/main/resources/spring/notify-consumer.xml
@@ -33,8 +33,8 @@
 
     <!-- generate proxy for the remote service, then demoService can be used in the same way as the
     local regular interface -->
-    <bean id="demoCallback" class="com.alibaba.dubbo.samples.notify.impl.NotifyImpl" />
-    <dubbo:reference id="demoService" check="false" interface="com.alibaba.dubbo.samples.notify.api.DemoService" version="1.0.0" group="cn">
+    <bean id="demoCallback" class="org.apache.dubbo.samples.notify.impl.NotifyImpl" />
+    <dubbo:reference id="demoService" check="false" interface="org.apache.dubbo.samples.notify.api.DemoService" version="1.0.0" group="cn">
         <dubbo:method name="sayHello" async="true" onreturn="demoCallback.onreturn" onthrow="demoCallback.onthrow"/>
     </dubbo:reference>
